小数点以下を排除

+++ KinectSkeletonTracking/MainWindow.xaml.cs
@@ -112,7 +112,7 @@ namespace KinectSkeletonTracking
                             var rollRotate = CalcRotate.Roll(orientation);
                         
                             // TODO:↑の角度の値から必要なものをソケット通信で送信する
-                            Debug.WriteLine(rollRotate.ToString());
+                            Debug.WriteLine(((int)rollRotate).ToString());
                         }
                     }
                 }
